Fix: 공통 컴포넌트화 #45
